requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
faiss-cpu==1.13.2
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-genai==1.73.1
greenlet==3.4.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-google-genai==4.2.2
langchain-groq==1.1.2
langchain-text-splitters==1.1.2
langsmith==0.7.32
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.19.0
numpy==2.4.4
orjson==3.11.8
packaging==26.1
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
pydeck==0.9.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
